Adding screenshot
Adding screenshot so people who can't log in can still see what it looks like - fixes issue #1

We unfortunately had to drop working on it midway through for something else but would very much like to pick it up and expand on the idea, with a more sophisticated implementation and as a continuously updating service (i.e. not just around specific elections, but maybe highliting them on the timeline).